FastAPI/requirements.txt
fastapi
pyJWT
passlib
pymongo
python-multipart